requirements.txt
fastapi>=0.100.1
pydantic>=2.1.1
uvicorn>=0.23.2
passlib>=1.7.4
python-jose>=3.3.0